feat: form inputs relations (#166)
An input can be linked to another input of the same form.
The dependent input will be activated or deactivated according the value of its dependency.

If the dependency is filled or true the input will be activated
If the dependency if empty or false the input will be deactivated